requirements.txt
streamlit
Pillow
numpy
stegano
